remove single-use pruner mixin
